Replace millis() async delays by a single timerqueue
This commit makes a start to get to a full async operation.
That means no or minimal blocking calls are made. A timerqueue
also makes it possible to implement timers inside the upcoming
rules.

The millis() method is pretty intertwined with some of the
functions of the current codebase. That means this rewrite
requires a more extensive review to check nothing is actually
broken.

Besides providing the original functionality just implemented
differently, this commit also add the improvement that the MQTT
and WiFi connection can be restored without having to reboot.

Another change is that the onS0Pulse interrupts now contains
more logic. I believe this adds such a marginal delay that
it doesn't effect normal operation.

The next step is to make the WiFiManager async.

----

This PR is the third of a series working towards implementing
rules.